[Silver III] Title: 구간 합 구하기 4, Time: 1068 ms, Memory: 59756 KB -BaekjoonHub

+++ 백준/Silver/11659. 구간 합 구하기 4/구간 합 구하기 4.java
@@ -3,14 +3,14 @@ import java.io.*;
 
 public class Main {
     public static void main(String[] args) throws IOException {
-        BufferedReader br =
+        BufferedReader br = 
             new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st =
             new StringTokenizer(br.readLine());
         
         int n = Integer.parseInt(st.nextToken());
         int m = Integer.parseInt(st.nextToken());
-        int[] a = new int[n+1];
+        long[] a = new long[n + 1];
         
         st = new StringTokenizer(br.readLine());
         for (int i = 1; i <= n; i++) {
@@ -21,7 +21,7 @@ public class Main {
             st = new StringTokenizer(br.readLine());
             int i = Integer.parseInt(st.nextToken());
             int j = Integer.parseInt(st.nextToken());
-            System.out.println(a[j] - a[i-1]);
+            System.out.println(a[j] - a[i - 1]);
         }
         br.close();
     }
